test: move footnote-defs scaling regression to test/specs/redos
Generated-by: blackhole-agent upstream-acceptance plane (autonomous stewardship mission)
